Extract ValueMapper and RowMapper to separate module
This helps keep the result parsing related logic isolated from client
module.
