+++ trino/client.py
@@ -34,8 +34,6 @@ The main interface is :class:`TrinoQuery`: ::
 """
 from __future__ import annotations
 
-import abc
-import base64
 import copy
 import functools
 import os
@@ -43,13 +41,12 @@ import random
 import re
 import threading
 import urllib.parse
-import uuid
 import warnings
 from dataclasses import dataclass
-from datetime import date, datetime, time, timedelta, timezone, tzinfo
-from decimal import Decimal
 from time import sleep
-from typing import Any, Dict, Generic, List, Optional, Tuple, TypeVar, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+from trino.mapper import RowMapper, RowMapperFactory
 
 try:
     from zoneinfo import ZoneInfo
@@ -57,7 +54,6 @@ except ModuleNotFoundError:
     from backports.zoneinfo import ZoneInfo
 
 import requests
-from dateutil import tz
 from tzlocal import get_localzone_name  # type: ignore
 
 import trino.logging
@@ -76,15 +72,6 @@ else:
     PROXIES = {}
 
 _HEADER_EXTRA_CREDENTIAL_KEY_REGEX = re.compile(r'^\S[^\s=]*$')
-
-T = TypeVar("T")
-
-PythonTemporalType = TypeVar("PythonTemporalType", bound=Union[time, datetime])
-POWERS_OF_TEN: Dict[int, Decimal] = {}
-for i in range(0, 13):
-    POWERS_OF_TEN[i] = Decimal(10 ** i)
-MAX_PYTHON_TEMPORAL_PRECISION_POWER = 6
-MAX_PYTHON_TEMPORAL_PRECISION = POWERS_OF_TEN[MAX_PYTHON_TEMPORAL_PRECISION_POWER]
 
 ROLE_PATTERN = re.compile(r"^ROLE\{(.*)\}$")
 
@@ -897,385 +884,3 @@ def _retry_with(handle_retry, handled_exceptions, conditions, max_attempts):
         return decorated
 
     return wrapper
-
-
-class ValueMapper(abc.ABC, Generic[T]):
-    @abc.abstractmethod
-    def map(self, value: Any) -> Optional[T]:
-        pass
-
-
-class NoOpValueMapper(ValueMapper[Any]):
-    def map(self, value) -> Optional[Any]:
-        return value
-
-
-class DecimalValueMapper(ValueMapper[Decimal]):
-    def map(self, value) -> Optional[Decimal]:
-        if value is None:
-            return None
-        return Decimal(value)
-
-
-class DoubleValueMapper(ValueMapper[float]):
-    def map(self, value) -> Optional[float]:
-        if value is None:
-            return None
-        if value == 'Infinity':
-            return float("inf")
-        if value == '-Infinity':
-            return float("-inf")
-        if value == 'NaN':
-            return float("nan")
-        return float(value)
-
-
-def _create_tzinfo(timezone_str: str) -> tzinfo:
-    if timezone_str.startswith("+") or timezone_str.startswith("-"):
-        hours = timezone_str[1:3]
-        minutes = timezone_str[4:6]
-        if timezone_str.startswith("-"):
-            return timezone(-timedelta(hours=int(hours), minutes=int(minutes)))
-        return timezone(timedelta(hours=int(hours), minutes=int(minutes)))
-    else:
-        return ZoneInfo(timezone_str)
-
-
-def _fraction_to_decimal(fractional_str: str) -> Decimal:
-    return Decimal(fractional_str or 0) / POWERS_OF_TEN[len(fractional_str)]
-
-
-class TemporalType(Generic[PythonTemporalType], metaclass=abc.ABCMeta):
-    def __init__(self, whole_python_temporal_value: PythonTemporalType, remaining_fractional_seconds: Decimal):
-        self._whole_python_temporal_value = whole_python_temporal_value
-        self._remaining_fractional_seconds = remaining_fractional_seconds
-
-    @abc.abstractmethod
-    def new_instance(self, value: PythonTemporalType, fraction: Decimal) -> TemporalType[PythonTemporalType]:
-        pass
-
-    @abc.abstractmethod
-    def to_python_type(self) -> PythonTemporalType:
-        pass
-
-    def round_to(self, precision: int) -> TemporalType:
-        """
-            Python datetime and time only support up to microsecond precision
-            In case the supplied value exceeds the specified precision,
-            the value needs to be rounded.
-        """
-        precision = min(precision, MAX_PYTHON_TEMPORAL_PRECISION_POWER)
-        remaining_fractional_seconds = self._remaining_fractional_seconds
-        digits = abs(remaining_fractional_seconds.as_tuple().exponent)
-        if digits > precision:
-            rounding_factor = POWERS_OF_TEN[precision]
-            rounded = remaining_fractional_seconds.quantize(Decimal(1 / rounding_factor))
-            if rounded == rounding_factor:
-                return self.new_instance(
-                    self.normalize(self.add_time_delta(timedelta(seconds=1))),
-                    Decimal(0)
-                )
-            return self.new_instance(self._whole_python_temporal_value, rounded)
-        return self
-
-    @abc.abstractmethod
-    def add_time_delta(self, time_delta: timedelta) -> PythonTemporalType:
-        """
-            This method shall be overriden to implement fraction arithmetics.
-        """
-        pass
-
-    def normalize(self, value: PythonTemporalType) -> PythonTemporalType:
-        """
-            If `add_time_delta` results in value crossing DST boundaries, this method should
-            return a normalized version of the value to account for it.
-        """
-        return value
-
-
-class Time(TemporalType[time]):
-    def new_instance(self, value: time, fraction: Decimal) -> TemporalType[time]:
-        return Time(value, fraction)
-
-    def to_python_type(self) -> time:
-        if self._remaining_fractional_seconds > 0:
-            time_delta = timedelta(microseconds=int(self._remaining_fractional_seconds * MAX_PYTHON_TEMPORAL_PRECISION))
-            return self.add_time_delta(time_delta)
-        return self._whole_python_temporal_value
-
-    def add_time_delta(self, time_delta: timedelta) -> time:
-        time_delta_added = datetime.combine(datetime(1, 1, 1), self._whole_python_temporal_value) + time_delta
-        return time_delta_added.time().replace(tzinfo=self._whole_python_temporal_value.tzinfo)
-
-
-class TimeWithTimeZone(Time, TemporalType[time]):
-    def new_instance(self, value: time, fraction: Decimal) -> TemporalType[time]:
-        return TimeWithTimeZone(value, fraction)
-
-
-class Timestamp(TemporalType[datetime]):
-    def new_instance(self, value: datetime, fraction: Decimal) -> Timestamp:
-        return Timestamp(value, fraction)
-
-    def to_python_type(self) -> datetime:
-        if self._remaining_fractional_seconds > 0:
-            time_delta = timedelta(microseconds=int(self._remaining_fractional_seconds * MAX_PYTHON_TEMPORAL_PRECISION))
-            return self.add_time_delta(time_delta)
-        return self._whole_python_temporal_value
-
-    def add_time_delta(self, time_delta: timedelta) -> datetime:
-        return self._whole_python_temporal_value + time_delta
-
-
-class TimestampWithTimeZone(Timestamp, TemporalType[datetime]):
-    def new_instance(self, value: datetime, fraction: Decimal) -> TimestampWithTimeZone:
-        return TimestampWithTimeZone(value, fraction)
-
-    def normalize(self, value: datetime) -> datetime:
-        if tz.datetime_ambiguous(value):
-            return self._whole_python_temporal_value.tzinfo.normalize(value)
-        return value
-
-
-class TimeValueMapper(ValueMapper[time]):
-    def __init__(self, precision):
-        self.time_default_size = 8  # size of 'HH:MM:SS'
-        self.precision = precision
-
-    def map(self, value) -> Optional[time]:
-        if value is None:
-            return None
-        whole_python_temporal_value = value[:self.time_default_size]
-        remaining_fractional_seconds = value[self.time_default_size + 1:]
-        return Time(
-            time.fromisoformat(whole_python_temporal_value),
-            _fraction_to_decimal(remaining_fractional_seconds)
-        ).round_to(self.precision).to_python_type()
-
-    def _add_second(self, time_value: time) -> time:
-        return (datetime.combine(datetime(1, 1, 1), time_value) + timedelta(seconds=1)).time()
-
-
-class TimeWithTimeZoneValueMapper(TimeValueMapper):
-    def map(self, value) -> Optional[time]:
-        if value is None:
-            return None
-        whole_python_temporal_value = value[:self.time_default_size]
-        remaining_fractional_seconds = value[self.time_default_size + 1:len(value) - 6]
-        timezone_part = value[len(value) - 6:]
-        return TimeWithTimeZone(
-            time.fromisoformat(whole_python_temporal_value).replace(tzinfo=_create_tzinfo(timezone_part)),
-            _fraction_to_decimal(remaining_fractional_seconds),
-        ).round_to(self.precision).to_python_type()
-
-
-class DateValueMapper(ValueMapper[date]):
-    def map(self, value) -> Optional[date]:
-        if value is None:
-            return None
-        return date.fromisoformat(value)
-
-
-class TimestampValueMapper(ValueMapper[datetime]):
-    def __init__(self, precision):
-        self.datetime_default_size = 19  # size of 'YYYY-MM-DD HH:MM:SS' (the datetime string up to the seconds)
-        self.precision = precision
-
-    def map(self, value) -> Optional[datetime]:
-        if value is None:
-            return None
-        whole_python_temporal_value = value[:self.datetime_default_size]
-        remaining_fractional_seconds = value[self.datetime_default_size + 1:]
-        return Timestamp(
-            datetime.fromisoformat(whole_python_temporal_value),
-            _fraction_to_decimal(remaining_fractional_seconds),
-        ).round_to(self.precision).to_python_type()
-
-
-class TimestampWithTimeZoneValueMapper(TimestampValueMapper):
-    def map(self, value) -> Optional[datetime]:
-        if value is None:
-            return None
-        datetime_with_fraction, timezone_part = value.rsplit(' ', 1)
-        whole_python_temporal_value = datetime_with_fraction[:self.datetime_default_size]
-        remaining_fractional_seconds = datetime_with_fraction[self.datetime_default_size + 1:]
-        return TimestampWithTimeZone(
-            datetime.fromisoformat(whole_python_temporal_value).replace(tzinfo=_create_tzinfo(timezone_part)),
-            _fraction_to_decimal(remaining_fractional_seconds),
-        ).round_to(self.precision).to_python_type()
-
-
-class BinaryValueMapper(ValueMapper[bytes]):
-    def map(self, value) -> Optional[bytes]:
-        if value is None:
-            return None
-        return base64.b64decode(value.encode("utf8"))
-
-
-class ArrayValueMapper(ValueMapper[List[Optional[Any]]]):
-    def __init__(self, mapper: ValueMapper[Any]):
-        self.mapper = mapper
-
-    def map(self, values: List[Any]) -> Optional[List[Any]]:
-        if values is None:
-            return None
-        return [self.mapper.map(value) for value in values]
-
-
-class NamedRowTuple(tuple):
-    """Custom tuple class as namedtuple doesn't support missing or duplicate names"""
-    def __new__(cls, values, names: List[str], types: List[str]):
-        return super().__new__(cls, values)
-
-    def __init__(self, values, names: List[str], types: List[str]):
-        self._names = names
-        # With names and types users can retrieve the name and Trino data type of a row
-        self.__annotations__ = dict()
-        self.__annotations__["names"] = names
-        self.__annotations__["types"] = types
-        elements: List[Any] = []
-        for name, value in zip(names, values):
-            if names.count(name) == 1:
-                setattr(self, name, value)
-                elements.append(f"{name}: {repr(value)}")
-            else:
-                elements.append(repr(value))
-        self._repr = "(" + ", ".join(elements) + ")"
-
-    def __getattr__(self, name):
-        if self._names.count(name):
-            raise ValueError("Ambiguous row field reference: " + name)
-
-    def __repr__(self):
-        return self._repr
-
-
-class RowValueMapper(ValueMapper[Tuple[Optional[Any], ...]]):
-    def __init__(self, mappers: List[ValueMapper[Any]], names: List[str], types: List[str]):
-        self.mappers = mappers
-        self.names = names
-        self.types = types
-
-    def map(self, values: List[Any]) -> Optional[Tuple[Optional[Any], ...]]:
-        if values is None:
-            return None
-        return NamedRowTuple(
-            list(self.mappers[index].map(value) for index, value in enumerate(values)),
-            self.names,
-            self.types
-        )
-
-
-class MapValueMapper(ValueMapper[Dict[Any, Optional[Any]]]):
-    def __init__(self, key_mapper: ValueMapper[Any], value_mapper: ValueMapper[Any]):
-        self.key_mapper = key_mapper
-        self.value_mapper = value_mapper
-
-    def map(self, values: Any) -> Optional[Dict[Any, Optional[Any]]]:
-        if values is None:
-            return None
-        return {
-            self.key_mapper.map(key): self.value_mapper.map(value) for key, value in values.items()
-        }
-
-
-class UuidValueMapper(ValueMapper[uuid.UUID]):
-    def map(self, value: Any) -> Optional[uuid.UUID]:
-        if value is None:
-            return None
-        return uuid.UUID(value)
-
-
-class NoOpRowMapper:
-    """
-    No-op RowMapper which does not perform any transformation
-    Used when legacy_primitive_types is False.
-    """
-
-    def map(self, rows):
-        return rows
-
-
-class RowMapperFactory:
-    """
-    Given the 'columns' result from Trino, generate a list of
-    lambda functions (one for each column) which will process a data value
-    and returns a RowMapper instance which will process rows of data
-    """
-    NO_OP_ROW_MAPPER = NoOpRowMapper()
-
-    def create(self, columns, legacy_primitive_types):
-        assert columns is not None
-
-        if not legacy_primitive_types:
-            return RowMapper([self._create_value_mapper(column['typeSignature']) for column in columns])
-        return RowMapperFactory.NO_OP_ROW_MAPPER
-
-    def _create_value_mapper(self, column) -> ValueMapper:
-        col_type = column['rawType']
-
-        if col_type == 'array':
-            value_mapper = self._create_value_mapper(column['arguments'][0]['value'])
-            return ArrayValueMapper(value_mapper)
-        elif col_type == 'row':
-            mappers = []
-            names = []
-            types = []
-            for arg in column['arguments']:
-                mappers.append(self._create_value_mapper(arg['value']['typeSignature']))
-                names.append(arg['value']['fieldName']['name'] if "fieldName" in arg['value'] else None)
-                types.append(arg['value']['typeSignature']['rawType'])
-            return RowValueMapper(mappers, names, types)
-        elif col_type == 'map':
-            key_mapper = self._create_value_mapper(column['arguments'][0]['value'])
-            value_mapper = self._create_value_mapper(column['arguments'][1]['value'])
-            return MapValueMapper(key_mapper, value_mapper)
-        elif col_type.startswith('decimal'):
-            return DecimalValueMapper()
-        elif col_type.startswith('double') or col_type.startswith('real'):
-            return DoubleValueMapper()
-        elif col_type.startswith('timestamp') and 'with time zone' in col_type:
-            return TimestampWithTimeZoneValueMapper(self._get_precision(column))
-        elif col_type.startswith('timestamp'):
-            return TimestampValueMapper(self._get_precision(column))
-        elif col_type.startswith('time') and 'with time zone' in col_type:
-            return TimeWithTimeZoneValueMapper(self._get_precision(column))
-        elif col_type.startswith('time'):
-            return TimeValueMapper(self._get_precision(column))
-        elif col_type == 'date':
-            return DateValueMapper()
-        elif col_type == 'varbinary':
-            return BinaryValueMapper()
-        elif col_type == 'uuid':
-            return UuidValueMapper()
-        else:
-            return NoOpValueMapper()
-
-    def _get_precision(self, column: Dict[str, Any]):
-        args = column['arguments']
-        if len(args) == 0:
-            return 3
-        return args[0]['value']
-
-
-class RowMapper:
-    """
-    Maps a row of data given a list of mapping functions
-    """
-    def __init__(self, columns):
-        self.columns = columns
-
-    def map(self, rows):
-        if len(self.columns) == 0:
-            return rows
-        return [self._map_row(row) for row in rows]
-
-    def _map_row(self, row):
-        return [self._map_value(value, self.columns[index]) for index, value in enumerate(row)]
-
-    def _map_value(self, value, value_mapper: ValueMapper[T]) -> Optional[T]:
-        try:
-            return value_mapper.map(value)
-        except ValueError as e:
-            error_str = f"Could not convert '{value}' into the associated python type"
-            raise trino.exceptions.TrinoDataError(error_str) from e
